+++ sdk/dapp-kit/src/utils/assertUnreachable.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Utility for compile-time exhaustiveness checking.
+ */
+export function assertUnreachable(value: never): never {
+	throw new Error(`ERROR! Encountered an unexpected value: ${value}`);
+}