+++ sdk/dapp-kit/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mysten/dapp-kit
 
+## 0.0.4
+
+### Patch Changes
+
+- @mysten/sui.js@0.41.2
+- @mysten/wallet-standard@0.7.2
+
 ## 0.0.3
 
 ### Patch Changes