+++ sdk/wallet-kit/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mysten/wallet-kit-core
 
+## 0.6.2
+
+### Patch Changes
+
+- @mysten/sui.js@0.41.2
+- @mysten/wallet-standard@0.7.2
+
 ## 0.6.1
 
 ### Patch Changes